+++ frontend/Forms/RECIPT VOUCHER.cs
@@ -64,7 +64,7 @@ namespace frontend.Forms
                 Print();
 
 
-                cmd = new SqlCommand("insert into recived values(N'" + nametb.Text + "',N'" + dateTP.Text + "',N'" + RfromCB.Text + "',N'" + categoryCB.Text + "','" + Int32.Parse(priceTb.Text) + "',N'" + priceIWCB.Text + "',N'" + discription.Text + "')", con);
+                cmd = new SqlCommand("insert into recived values(N'" + nametb.Text + "',N'" + dateTP.Text + "',N'" + RfromCB.Text + "',N'" + categoryCB.Text + "','" + Int64.Parse(priceTb.Text) + "',N'" + priceIWCB.Text + "',N'" + discription.Text + "')", con);
                 try
                 {
                     con.Open();
